pyproject.toml
[project]
name = "vk-wall-parser"
version = "0.1.0"
description = "Short script to parse all posts from VK group using vk_api"
authors = ["<yakovlevs01@yandex.ru>"]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "certifi==2023.11.17",
    "charset-normalizer==3.3.2",
    "idna==3.6",
    "python-dotenv==1.0.0",
    "requests==2.31.0",
    "urllib3==2.1.0",
    "vk-api==11.9.9",
]


[dependency-groups]
dev = [
    "ruff>=0.11.2",
]


[tool.ruff]
line-length = 120
fix = true
show-fixes = true

[tool.ruff.lint]
select = ["ALL"]
ignore = ["EM", "FBT", "TRY003", "S101", "D", "FA", "T201", "INP001", "COM812", "RUF001", "RUF002", "RUF003", "N806", "B018"]
pydocstyle.convention = "google"
extend-safe-fixes = []
